nimbus_test/requirements.txt
django
psycopg2-binary
fpdf
